move: Notice 클래스 패키지 생성 및 이동
